fix: move types to separate file too noisy working in index.ts
